updated for public portfolio
I update the content and formatting of the README along with the Project reflection to make them more accurate, readable, and comprehensive. 

I also modified the program so that L, T, and P inputs could only be read every half second. 

added the dependencyTree.txt to make it easier for others to more easily get this project working on their machine.

added demonstration video.